backend/requirements.txt

fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.3
python-dotenv==1.0.0
supabase==1.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.1
gunicorn==20.1.0
pydantic-settings==2.0.3
